pyproject.toml
[project]
name = "rentcast-mcp-server"
version = "0.1.0"
description = "MCP server for RentCast API integration"
license = {text = "MIT"}
readme = "README.md"
authors = [
    { name = "Rob", email = "rob@example.com" }
]
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.3.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/rentcast-mcp-server"
"Bug Tracker" = "https://github.com/yourusername/rentcast-mcp-server/issues"

[project.scripts]
rentcast-mcp = "rentcast_mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rentcast_mcp_server"]

[tool.hatch.build]
only-packages = true

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "B", "W"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 